Enable TLS encryption of internal API
Access to OpenStack APIs and the dashboard is now encrypted, terminated by
HAProxy. The combined certificate and key is encrypted using Ansible vault in
this repository. The private root CA used to sign the certificate is added to
the etc/kayobe/kolla/certificates/ca/ directory, so it will be copied to the
trust store of every container.