Enable separate external API with TLS
Accessible via https://the.greenedge.cloud.

The external API requires a separate routing table to avoid asymmetric routing.
This is applied via kayobe. There is currently an issue in that the network
configuration fails if the VIP is active on the interface during the following
command:

  kayobe overcloud host configure

This needs to be fixed, but can be worked around by passing --skip-tags
network.