Fix API VIP failures
Using keepalived to apply IP rules and routes dynamically seems to cause the
VIP to fail. The trigger appears to be network interfaces being created in
network namespaces for DHCP agents etc, causing keepalived to enter a FAULT
state.

This change addresses the issue by switching to static routes applied by
Kayobe. This depends on patches in Kayobe and the MichaelRigart.interfaces
Ansible role.