Add templating back to prometheus.yml config
The prometheus.yml configuration file is currently hard-coded to include IPs of
targets. This is not ideal, since it is not updated when hosts are added or
removed from the cluster.

This change syncs the template with the one in Kolla-Ansible, adding the custom
scrape configs at the end.

The generated config has been compared before and after this change.