Tag built Kolla with the date & time
This allows for more fine grained control over where and when images are
deployed. It also allows images to be rolled back.

kolla_tag defines the tag used when building.

The tags added to globals.yml define which tags are deployed for each service.
Initially they are all set to train-centos8, which is the tag currently in use.