Assign static IPs to Ceph nodes on public network
Keepalived does not always function correctly when the interface does not have
a static IP address. Bouncing the interface can leave it without a VIP, and
restarting keepalived can leave the interface without static routes.

Also, use a separate public2 network for RadosGW loadbalancer.

In production, this network uses a different subnet than the OpenStack public
API. Kayobe will fail when allocating IPs if they are not in the correct
subnet, so we must make a second public network for the RadosGW loadbalancer.