Don't configure public network on external interface
I originally did this so that I could run tempest from outside of the VM and I
needed the service catalog to use the external IP, but that plan had a floor...
You still need to be able to access VMs. Instead, I've switched to running
tempest via ansible so that I can run it one the remote host. Using an
isolated bridge has the advantage that the tempest_openrc will remain unchanged
between all in ones so that we don't need a different openrc file per
environment (the public IP is fixed).