Brings in growroot and swap playbooks from stackhpc-kayobe-config
Brings in the growroot and swap playbooks from stackhpc-kayobe-config.  I've
changed the namespacing of lvm variables to scientific-openstack so that we can
make these match our overcloud-dib configuration.

Tests that they work on the AIO. Prior to this change, we were running
out of space when doing an upgrade. The AIO has been modified to grow
the var partition to 40G to accomodate both sets of images.