Adds a Rocky 8 AIO test (#194)
* Adds Rocky 8 pulp repositories to dnf.yml (#152)

* Adds Rocky 8 pulp repositories to dnf.yml

This configures the hosts to use the mirrors in the local
pulp. Prior to this change kayobe would attempt to configure
CentOS stream mirrors on a Rocky host.

* Use facts to pick dnf repositories

* Adds Rocky 8 repository config for local pulp (#146)

* Adds Rocky 8 repository config for local pulp

We support Rocky 8 as a host OS with os_distribution == 'rocky'. This
adds the configuration necessary to use the RPM snapshots from ark.

* Abstract over version number v1

Unsure if it would be better to:

1) Update the remote on the repository
2) Change the publication to point to a new repository and keep
repositories for every minor version

Advantages of 1:

- Single repository in pulp

Advantages of 2:

- Less chance of conflicts (although I think an old repo is just an old snapshot)

* Fixups

* Address code review comment

Upadting variable name after feedback

* Update docker baseurl to work with Rocky 8 (#153)

This is the equivalent change of:

https://review.opendev.org/c/openstack/ansible-collection-kolla/+/831642

* Adds a Rocky 8 AIO test

* Use ansible fork with Rocky 8 support

* Install fork of ansible unconditionally

The kayobe docker image fron kayobe-automation was being built
without Rocky Linux support. This is because os_distribution was
not being set add build time.

* Workaround lack of ovs-vsctl on the host.

* Switch back to stackhpc/wallaby

The PR has now been merged.

* centos user has changed to cloud-user

The images on smslab are now based on centos-minimal.

* Add os-distribution to tempest artifact name

This allows you to differentiate the artifacts across the distros.

* Remove bootstrap override

This is no longer necessary now the centos image also uses cloud-user.

* Revert volume_size to 100GB

200G was only necessary for testing the DIB images from kayobe-config
with the default LVM logical volume sizes.

* Remove configure-local-networking