Add SMART Monitoring with dash and alerts
Enabled Textfile collector in node exporter in kolla/globals.yml

Added smartmon script as is from the prometheus-community github and then
removed NVME support from this script in favour of using the nvme-cli script,
which has also been added in. This is because the nvme-cli script provides
better metrics than the smartmon script does. The script also adds the serial
number of the disk as a label to all SMART metrics.

Added a Kayobe custom playbook to easily deploy the script and associated
cron job. This playbook installs smartmontool and nvmecli then copies these
over to the hosts and sets up a cronjob which runs the scripts and stores
the metrics in the docker volume for node exporter. The playbook changes
the way the metrics are saved to a file by making use of the mv command
as it is atomic. This was needed as at times prometheus would read a
partially completed file.

Added a prometheus alert to alert when a drive is reported as not healthy
for more than 10 minutes.

Added a Grafana dashboard to display the number of healthy and unhealthy
drives reported in prometheus.