habrok: EVPN: increase MTU of BGP peer VLAN IP interface
This seems to be the missing config from the Dell reference design that
prevents EVPN BGP session flapping.

The EVPN BGP traffic may be routed directly to a leaf, or via the leaf
peer, then the inter-switch link. In the second case, it uses this VLAN
IP. It needs to have the same MTU as the fabric links.