Update the hardware overview dashboard
Adds the temperature of drives as a column to the table.
Adds a timeseries to track the history of the temperature.
Introduces a variable dropdown picker for nodes so that only information
about disks on a specific node is shown.

Change-Id: Ibff3037865e1891d8e07a945dbeedf08e1f8c0d2
(cherry picked from commit 4999f9c910204090d05b3338c5c60d73cf07b2dc)