Refactor tag overrides
This refactor includes two changes

The first sets kolla_base_distro according to ansible facts, and is
resolved by kolla-ansible on a host by host basis.

The second reformats the existing structure for override tags,
again allowing kolla-ansible to resolve them on a host by host basis.

The result of these two changes is that clouds with with mixed host
distributions can be deployed.

This is relevant when performing a rolling migration between two
distributions.