Add a custom playbook to fix OVN chassis priorities
Sometimes, typically after restarting OVN services, the priorities of entries
in the ha_chassis and gateway_chassis tables in the OVN northbound database
can become misaligned. This results in broken routing for external (bare
metal/SR-IOV) ports.

This playbook can be used to fix the issue by realigning the priorities of
the table entries. It does so by assigning the highest priority to the
"first" (sorted alphabetically) OVN NB DB host. This results in all gateways
being scheduled to a single host, but is less complicated than trying to
balance them (and it's also not clear to me how to map between individual
ha_chassis and gateway_chassis entries).

The playbook can be run as follows:
kayobe playbook run $KAYOBE_CONFIG_PATH/ansible/ovn-fix-chassis-priorities.yml

If the 'controllers' group does not align with the group used to deploy the
OVN NB DB, this can be overridden by passing the following:
'-e ovn_nb_db_group=some_other_group'