DNM: CloudKitty  for fresh OpenSearch deployments
This container tag brings in the fixes from this PR:
https://github.com/stackhpc/cloudkitty/pull/57 which
allows CloudKitty with the elasticsearch backend to be
used with fresh OpenSearch. i.e for new deployments
(Habrok, AZ Moldnal, etc.)

This PR should not be merged until we are happy with
using the same patch for existing deployments after an
ES->OS migration. (Or ideally, never at all if we want to
integrate the proper opensearch v2 storage driver)