CI: Skip aio jobs based on changed files
We don't always want to run all aio jobs - it can be wasteful of CI
resources and slows down merging changes.

This change uses the dorny/paths-filter action to detect changed files
and skip jobs as appropriate.

Note that we can't use the workflow-level paths attribute since this
would skip the workflow entirely, and would prevent us from making the
aio jobs required to pass (a skip counts as a pass).