Ubuntu: add base repos to Kolla container image earlier
Previously when building Kolla container images for Ubuntu it was common
to get a dependency error during apt package installation, such as:

  The following packages have unmet dependencies:
    libc6-dev : Depends: libc6 (= 2.35-0ubuntu3.1) but 2.35-0ubuntu3.4 is to be installed

This is caused by the packages in the Ubuntu base image being newer than
those in the release train repository snapshot.

This change fixes the issue by switching to the release train
repositories earlier in the build process, before the first apt command
runs.

It's possible that we may also need to pin the Ubuntu base container
image, but that would be another thing to maintain, so let's try without
for now since this approach is necessary and may be sufficient.