Fix Grafana HAProxy dashboard (again)
The previous commit to this file [1] added port numbers to the instance
label regex. This is compatible with the default Prometheus targets
generated by Kolla Ansible, which look like this:

    192.168.1.1:9101

However, when using a non-default instance label [2], the port number is
absent, which breaks the dashboard (all panels are empty).

Modify the regex to make the port number optional, which should support
all possible instance labels.

Note: I first tried using `$host:([0-9]+)?` but it does not appear to be
supported by Prometheus, which uses RE2 [3].

[1] https://github.com/stackhpc/stackhpc-kayobe-config/commit/eceee825432748c48035ae1f3b3289a14c7373c0
[2] https://docs.openstack.org/kolla-ansible/latest/reference/logging-and-monitoring/prometheus-guide.html#metric-instance-labels
[3] https://github.com/google/re2/wiki/Syntax