RL9: Set SELinux to permissive mode in the standard overcloud host image
From the Zed release, Kayobe changes its default SELinux mode from
disabled to permissive. Changing SELinux from disabled to permissive
mode requires a reboot, so it's best to include the correct mode in the
image.

This change is proposed to the Yoga branch because that is the point of
the RL9 migration, when it is most convenient to make this change.
Doing so requires a backport of the Kayobe patch supporting SELinux
modes, and for Kayobe config to set selinux_mode to permissive (we would
not change the default config in a backport).