Add alertmanager rules for Docker
Docker has a builtin prometheus exporter that we currently don't have
enabled. This change adds alerts for stopped/paused containers and
failed healthchecks. This patch requires changes to docker's
configuration to export the metrics, and prometheus to consume them.
This means that Kayobe and Kolla-Ansible should both be updated to their
latest versions.