CI: Run container image build job on ARC cluster
The Actions Runner Controller (ARC) cluster runs GitHub Actions jobs on
a Kubernetes cluster. This allows us to dynamically provision resources
for our CI workflows.

This change switches the container image build jobs to use the ARC
cluster. The job now runs in a container, with access to a Docker in
Docker container for building images. This allows us to skip the host
configuration part of the workflow.

See https://github.com/stackhpc/ARC-Installer/