Add sufficient detail the the app creds fixes
Specify that app creds in Yoga are affected by the lack of implicit
roles, so no assumptions are made that a system is safe if it is a
relatively new deployment.

Add a note to the issue of implicit reader roles seen at Bristol Uni.

Add instructions on how to add the reader role to existing app creds,
taken from Slack conversations