feat!: change when `pre-commit` hooks are installed
Initally `pre-commit` hooks were installed due the presense of a hook
for `control host bootstrap` and when `enable_pre_commit_hooks` was set
to `true`. However now `pre-commit` hooks are only installed when the
hook is present or the playbook is called, no requirement for an
`Ansible` conditional to be true.