os_capacity: Add tags to playbook, update vault docs
Previously the first deployment of a system with a Vault CA for internal
TLS and os_capacity enabled would fail when deploying HAProxy.
os_capacity deployment requires admin-openrc.sh to exist, but because of
the use of -kt haproxy the post-deploy tasks that create it will be
skipped.

This change fixes the issue by adding an os_capacity tag to the relevant
plays, and updating the Vault docs to skip the new tag when deploying
HAProxy.