fix: LVM expects a dictionary for `seed-hypervisor`
```
seed_hypervisor_lvm_groups:
  - "{{ seed_hypervisor_lvm_groups_with_data }}"
  - "{{ lvm_group_vg_rl_vm3 }}"
```
Is incorrect as the value of `seed_hypervisor_lvm_groups_with_data` is a
list and not a dictionary. This is fine if the use it a single item for
` seed_hypervisor_lvm_groups` but not with multiple groups.

Solution has been to use `first` filter to extract dictionary from the
list.