fix: subnet and gateway for `external_intranet`
During the deployment of the network configuration for the
`seed-hypervisor` connection was lost and required assistence from the
internal team to regain access.

Once access was restored the network configuration for the
`external_intranet` had changed from `/32` to `/24` which both differ from
the `/23` as mentioned in the reference architecture. To avoid any
further delay or disruption I opted to keep the existing configuration for
the `external_intranet` network.