CI: Register overcloud host image as shared and add stackhpc-dev
In order for developers to be able to use overcloud images for testing,
they must be available in the stackhpc-dev project on Leafcloud. To do
this we can make the image shared, and add stackhpc-dev as a member.
Unfortunately someone in the stackhpc-dev project must also accept the
image membership before it becomes active. It's not possible to do this
with the existing app creds which are scoped to the stackhpc-ci project.
Instead we provide instructions on how to do it as a debug message.