Run StackHPC OpenStack tests in aio CI
This change adds a new playbook, stackhpc-openstack-tests.yml, that uses
the new stackhpc-openstack-tests repository to perform testing of the
deployment that is not covered by Rally or Tempest. Currently this
includes Prometheus and OpenSearch.

The all-in-one CI test has been updated to run this new test suite, and
the results are included in the test results artifact.