CI: Separate monitoring vs host test execution in SOT
There are currently two types of tests in StackHPC OpenStack Tests:
service-level tests such as those for monitoring, and host-level tests
such as those for Docker and SELinux. We should execute the
service-level tests from a single remote host, while the host-level
tests should be executed on each of the overcloud hosts.

This means that we now get several results files, so it becomes
important to provide a summary of failures.