Actually check key type
Don't assume that users won't lie to you. I actually hit this when
following the terraform multinode README. My version of openssh will
generate ed25519 keys by default and the example command in the README
names this id_rsa[1]. This of course ends up locking you out of the
system.

[1] https://github.com/stackhpc/terraform-kayobe-multinode/blob/a3e72a90c8ce835cc2bb4cbda83e79b0c32afdde/README.rst#L73