kolla-images.py: Fix image to container exceptions using check-image-map
We recently had an issue where images were built for ironic, but this
did not include dnsmasq due to not matching the regex. The 'ironic' tag
was updated in kolla-image-tags.yml, which matches ironic_dnsmasq
container, but no such dnsmasq image existed. This issue would have been
caught the check-tags workflow with the correct mapping added in this
change.