Increase size of connection tracking table
We have seen occurences of Cinder or Glance activity filling up the
conntrack table. This has been seen on controllers and storage nodes
(Ceph). We have also seen large amounts of traffic to an Octavia amphora
causing the conntrack table of a compute host to fill up.

Quadruple the maximum size of the table on these hosts (the default size
is 262,144 for hosts with more than 4 GiB of memory [1]).

Monitoring hosts are left unchanged since they are believed to be
unaffected.

[1] https://www.kernel.org/doc/Documentation/networking/nf_conntrack-sysctl.txt