Add ansible-lint CI
This commit includes a properly configured ansible-lint CI job and a
large amount of changes to the existing playbooks so that the new CI
passes. Some fixes were applied automatically with the --fix argument,
many were made manually. There is some risk that the changes have
altered the behaviour of the playbooks.