Update rabbitmq-reset for Epoxy
With the introduction of oslo.messaging Queue Manager, we need to clear
shared memory when wiping queues from RabbitMQ. This is because
oslo.messaging assumes that a queue with an entry under shared memory
exists in the RabbitMQ cluster.

With the introduction of durable queues for reply/fanout queues, we need
stop the services before wiping queues from RabbitMQ. This is because
fanout queues in particular are not just created on container startup,
so services may attempt to use them after the RabbitMQ state has been
reset.

Removes the tags as it no longer makes sense to only run part of this
playbook.