Redfish Exporter v2.x (#1549)
* Configuration for the StackHPC fork of Redfish Exporter

* Update redfish dashboard

- Compatability with Dell servers
- Added health round up panels
- Add variables for different groups e.g overcloud vs compute

* Increase query lookback delta

Prometheus will regard metrics collected over this perioid as stale as
such they won't show up in grafana.

* Add raw tags

* Re-tweak Redfish dashboard for Lenovo

* Restore old UUID of the redfish dashboard

The redfish dashboard was updated to use a new UUID, however this caused
errors in the Grafana logs and Grafana didn't update the dashboard.
Reverting this to the old UUID fixes the issue.

* Make query lookback delta setting smarter

* Actually set lookback delta

* Adds openstack exporter scrape interval

* Add some raw tags

* Slightly less offensive variant on raw tags

* Refresh release note

* Update dashboard

* Apply suggestions from code review

Co-authored-by: Alex-Welsh <112560678+Alex-Welsh@users.noreply.github.com>

* Bump to redfish exporter v2.1.1-stackhpc

* Temporarily remove redfish log collection

This isn't working yet (takes far too long to scrape),let's
split it out into a separate patch.

Also removed the query lookback delta config, as this was
working around the much longer scrape times.

---------

Co-authored-by: Matt Anson <matta@stackhpc.com>
Co-authored-by: technowhizz <7688823+technowhizz@users.noreply.github.com>
Co-authored-by: Matt Crees <mattc@stackhpc.com>
Co-authored-by: Alex-Welsh <112560678+Alex-Welsh@users.noreply.github.com>