fix: use `rabbitmq` length for `RabbitMQNodeDown` (#1579)
The `RabbitMQNodeDown` made the assumption that all deployments involve
only three RabbitMQ nodes. However, this is not always the case as we
do support deployments with a single node or more than three.

Before this would have caused false alerts in deployments with a single
RabbitMQ node. Whilst also concealing alerts in deployments with more
than three nodes.

Co-authored-by: Matt Crees <mattc@stackhpc.com>