Clone Kayobe into a tempdir when doing version checks
Cloning Kayobe as part of check-kayobe-version.yml
would fail with POSIX permission errors when it was run by multiple users, because the clone destination
directory is a) always the same and b) never cleaned up.

Clone Kayobe into a unique directory each time using
ansible.builtin.tempfile and clean up the directory tree
when it is no longer needed. This ensures that each
user that runs check-kayobe-version.yml gets a
Kayobe checkout to a unique path, and not one that
already exists and is owned by someone else.